pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "mcp-gcp-proxy"
version = "0.1.2"
description = "MCP stdio to HTTP proxy with Google impersonation auth backends"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-auth>=2.35.0",
  "httpx>=0.28.0",
  "httpx-retries>=0.4.0",
  "mcp>=1.12.0",
  "requests>=2.32.0",
]

[project.scripts]
mcp-googleapis-proxy = "mcp_gcp_proxy.cli:googleapis_main"
mcp-cloudrun-proxy = "mcp_gcp_proxy.cli:cloudrun_main"

[project.optional-dependencies]
dev = [
  "mypy>=1.15.0",
  "pytest>=8.3.0",
  "ruff>=0.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gcp_proxy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N", "SIM", "C4", "ARG", "W"]
ignore = ["ARG001"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
files = ["src/mcp_gcp_proxy"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
